# src/main/resources/openapi-spec.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/tasks/{id}":{"get":{"tags":["task-controller"],"operationId":"getTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}},"put":{"tags":["task-controller"],"operationId":"putTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}},"delete":{"tags":["task-controller"],"operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}},"patch":{"tags":["task-controller"],"operationId":"markAsDone","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"boolean"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/tasks":{"get":{"tags":["task-controller"],"operationId":"getTasks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}},"post":{"tags":["task-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/registration":{"post":{"tags":["Auth"],"operationId":"createNewUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationUserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/confirm-reset":{"post":{"tags":["Auth"],"operationId":"confirmResetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeUserPasswordDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/authentication":{"post":{"tags":["Auth"],"operationId":"createAuthToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/unsecured":{"get":{"tags":["hello-controller"],"operationId":"unsecured","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/secured":{"get":{"tags":["hello-controller"],"operationId":"sayHello","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/reset-password":{"get":{"tags":["Auth"],"operationId":"resetPassword","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/current":{"get":{"tags":["hello-controller"],"operationId":"currentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/admin":{"get":{"tags":["hello-controller"],"operationId":"admin","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Task":{"required":["date"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","description":"Test description","format":"date"},"description":{"type":"string"},"done":{"type":"boolean"}},"description":"Task description"},"RegistrationUserDto":{"type":"object","properties":{"username":{"type":"string"},"mail":{"type":"string"},"password":{"type":"string"}}},"ChangeUserPasswordDto":{"type":"object","properties":{"token":{"type":"string"},"password":{"type":"string"}}},"JwtRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}}